# Conflicts:
#	clients/pipeline-service/pms-sdk-core/src/main/java/io/harness/plancreator/stages/parallel/ParallelPlanCreator.java

+++ 440-connector-nextgen/src/main/java/io/harness/connector/impl/DefaultConnectorServiceImpl.java
@@ -11,6 +11,7 @@ import static io.harness.NGConstants.CONNECTOR_HEARTBEAT_LOG_PREFIX;
 import static io.harness.NGConstants.CONNECTOR_STRING;
 import static io.harness.NGConstants.HARNESS_SECRET_MANAGER_IDENTIFIER;
 import static io.harness.beans.FeatureName.NG_SETTINGS;
+import static io.harness.beans.FeatureName.PL_FORCE_DELETE_CONNECTOR_SECRET;
 import static io.harness.connector.ConnectivityStatus.FAILURE;
 import static io.harness.connector.ConnectivityStatus.UNKNOWN;
 import static io.harness.data.structure.EmptyPredicate.isEmpty;
@@ -19,7 +20,6 @@ import static io.harness.errorhandling.NGErrorHelper.DEFAULT_ERROR_SUMMARY;
 import static io.harness.exception.WingsException.USER;
 import static io.harness.git.model.ChangeType.ADD;
 import static io.harness.utils.PageUtils.getPageRequest;
-import static io.harness.utils.RestCallToNGManagerClientUtils.execute;
 
 import static java.lang.Boolean.parseBoolean;
 import static java.lang.String.format;
@@ -54,6 +54,7 @@ import io.harness.connector.entities.Connector;
 import io.harness.connector.entities.Connector.ConnectorKeys;
 import io.harness.connector.events.ConnectorCreateEvent;
 import io.harness.connector.events.ConnectorDeleteEvent;
+import io.harness.connector.events.ConnectorForceDeleteEvent;
 import io.harness.connector.events.ConnectorUpdateEvent;
 import io.harness.connector.helper.CatalogueHelper;
 import io.harness.connector.helper.HarnessManagedConnectorHelper;
@@ -77,7 +78,6 @@ import io.harness.delegate.beans.connector.scm.genericgitconnector.GitConfigDTO;
 import io.harness.delegate.beans.connector.scm.github.GithubConnectorDTO;
 import io.harness.delegate.beans.connector.scm.gitlab.GitlabConnectorDTO;
 import io.harness.encryption.SecretRefData;
-import io.harness.entitysetupusageclient.remote.EntitySetupUsageClient;
 import io.harness.errorhandling.NGErrorHelper;
 import io.harness.eventsframework.schemas.entity.EntityDetailProtoDTO;
 import io.harness.eventsframework.schemas.entity.IdentifierRefProtoDTO;
@@ -107,6 +107,7 @@ import io.harness.ng.core.NGAccess;
 import io.harness.ng.core.dto.ErrorDetail;
 import io.harness.ng.core.entities.Organization;
 import io.harness.ng.core.entities.Project;
+import io.harness.ng.core.entitysetupusage.service.EntitySetupUsageService;
 import io.harness.ng.core.services.OrganizationService;
 import io.harness.ng.core.services.ProjectService;
 import io.harness.ngsettings.SettingIdentifiers;
@@ -160,7 +161,6 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
   private final CatalogueHelper catalogueHelper;
   private final ProjectService projectService;
   private final OrganizationService organizationService;
-  EntitySetupUsageClient entitySetupUsageClient;
   ConnectorStatisticsHelper connectorStatisticsHelper;
   NGSettingsClient settingsClient;
   private NGErrorHelper ngErrorHelper;
@@ -173,6 +173,7 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
   GitSyncSdkService gitSyncSdkService;
   OutboxService outboxService;
   YamlGitConfigClient yamlGitConfigClient;
+  EntitySetupUsageService entitySetupUsageService;
 
   @Override
   public Optional<ConnectorResponseDTO> get(
@@ -719,13 +720,14 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
   }
 
   @Override
-  public boolean delete(
-      String accountIdentifier, String orgIdentifier, String projectIdentifier, String connectorIdentifier) {
-    return delete(accountIdentifier, orgIdentifier, projectIdentifier, connectorIdentifier, ChangeType.DELETE);
+  public boolean delete(String accountIdentifier, String orgIdentifier, String projectIdentifier,
+      String connectorIdentifier, boolean forceDelete) {
+    return delete(
+        accountIdentifier, orgIdentifier, projectIdentifier, connectorIdentifier, ChangeType.DELETE, forceDelete);
   }
 
   public boolean delete(String accountIdentifier, String orgIdentifier, String projectIdentifier,
-      String connectorIdentifier, ChangeType changeType) {
+      String connectorIdentifier, ChangeType changeType, boolean forceDelete) {
     Optional<Connector> existingConnectorOptional =
         getInternal(accountIdentifier, orgIdentifier, projectIdentifier, connectorIdentifier);
     if (!existingConnectorOptional.isPresent()) {
@@ -734,15 +736,40 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
     }
     Connector existingConnector = existingConnectorOptional.get();
     ConnectorResponseDTO connectorDTO = connectorMapper.writeDTO(existingConnector);
-    checkThatTheConnectorIsNotUsedByOthers(existingConnector);
-    connectorEntityReferenceHelper.deleteConnectorEntityReferenceWhenConnectorGetsDeleted(
-        connectorDTO.getConnector(), accountIdentifier);
+    if (forceDelete && !isForceDeleteEnabled(accountIdentifier)) {
+      throw new InvalidRequestException(
+          format(
+              "Parameter forcedDelete cannot be true. Force Delete is not enabled for account [%s]", accountIdentifier),
+          USER);
+    }
+
+    if (!forceDelete) {
+      checkThatTheConnectorIsNotUsedByOthers(existingConnector);
+    }
+    try {
+      connectorEntityReferenceHelper.deleteConnectorEntityReferenceWhenConnectorGetsDeleted(
+          connectorDTO.getConnector(), accountIdentifier);
+    } catch (Exception e) {
+      if (forceDelete) {
+        log.warn(
+            "Exception occured while deleting references of connector {} with accountId {}, orgId {}, projectId {} - {} ",
+            connectorIdentifier, accountIdentifier, orgIdentifier, projectIdentifier, e);
+      } else {
+        throw e;
+      }
+    }
     existingConnector.setDeleted(true);
     Supplier<OutboxEvent> supplier = null;
     if (!gitSyncSdkService.isGitSyncEnabled(accountIdentifier, orgIdentifier, projectIdentifier)) {
-      supplier = ()
-          -> outboxService.save(
-              new ConnectorDeleteEvent(accountIdentifier, connectorMapper.writeDTO(existingConnector).getConnector()));
+      if (forceDelete) {
+        supplier = ()
+            -> outboxService.save(new ConnectorForceDeleteEvent(
+                accountIdentifier, connectorMapper.writeDTO(existingConnector).getConnector()));
+      } else {
+        supplier = ()
+            -> outboxService.save(new ConnectorDeleteEvent(
+                accountIdentifier, connectorMapper.writeDTO(existingConnector).getConnector()));
+      }
     }
 
     connectorRepository.delete(existingConnector, null, changeType, supplier);
@@ -767,8 +794,8 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
                                       .build();
     String referredEntityFQN = identifierRef.getFullyQualifiedName();
     try {
-      isEntityReferenced = execute(entitySetupUsageClient.isEntityReferenced(
-          connector.getAccountIdentifier(), referredEntityFQN, EntityType.CONNECTORS));
+      isEntityReferenced = entitySetupUsageService.isEntityReferenced(
+          connector.getAccountIdentifier(), referredEntityFQN, EntityType.CONNECTORS);
     } catch (Exception ex) {
       log.info("Encountered exception while requesting the Entity Reference records of [{}], with exception",
           connector.getIdentifier(), ex);
@@ -1167,9 +1194,8 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
 
   private Boolean isBuiltInSMDisabled(String accountIdentifier) {
     Boolean isBuiltInSMDisabled = false;
-    boolean isNgSettingsEnabled =
-        CGRestUtils.getResponse(accountClient.isFeatureFlagEnabled(NG_SETTINGS.name(), accountIdentifier));
-    if (isNgSettingsEnabled) {
+
+    if (isNgSettingsFFEnabled(accountIdentifier)) {
       isBuiltInSMDisabled = parseBoolean(
           NGRestUtils
               .getResponse(settingsClient.getSetting(
@@ -1177,5 +1203,31 @@ public class DefaultConnectorServiceImpl implements ConnectorService {
               .getValue());
     }
     return isBuiltInSMDisabled;
+  }
+
+  private boolean isForceDeleteEnabled(String accountIdentifier) {
+    boolean isForceDeleteFFEnabled = isForceDeleteFFEnabled(accountIdentifier);
+    boolean isForceDeleteEnabledBySettings =
+        isNgSettingsFFEnabled(accountIdentifier) && isForceDeleteFFEnabledViaSettings(accountIdentifier);
+    return isForceDeleteFFEnabled && isForceDeleteEnabledBySettings;
+  }
+
+  @VisibleForTesting
+  protected boolean isForceDeleteFFEnabledViaSettings(String accountIdentifier) {
+    return parseBoolean(NGRestUtils
+                            .getResponse(settingsClient.getSetting(
+                                SettingIdentifiers.ENABLE_FORCE_DELETE, accountIdentifier, null, null))
+                            .getValue());
+  }
+
+  @VisibleForTesting
+  protected boolean isForceDeleteFFEnabled(String accountIdentifier) {
+    return CGRestUtils.getResponse(
+        accountClient.isFeatureFlagEnabled(PL_FORCE_DELETE_CONNECTOR_SECRET.name(), accountIdentifier));
+  }
+
+  @VisibleForTesting
+  protected boolean isNgSettingsFFEnabled(String accountIdentifier) {
+    return CGRestUtils.getResponse(accountClient.isFeatureFlagEnabled(NG_SETTINGS.name(), accountIdentifier));
   }
 }
